requirements/testing.txt
-r base.txt

django-webtest
beautifulsoup4
factory_boy
lxml
